+++ ray_tracing/sphere.hpp
@@ -19,7 +19,7 @@ public:
     , m_radius(radius)
   {}
 
-  uint32_t GetType() const override { return kHitableObjectSphereType; }
+  uint8_t GetType() const override { return kHitableObjectSphereType; }
 
   glm::vec3 const GetCenter() const { return m_center; }
   float GetRadius() const { return m_radius; }
